requirements.txt
alembic==1.4.0
blinker==1.4
Click==7.0
decorator==4.4.1
Flask==1.1.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WhooshAlchemyPlus==0.7.4
Flask-WTF==0.14.3
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
marshmallow==3.5.0
marshmallow-sqlalchemy==0.22.2
passlib==1.7.2
phonenumbers==8.11.4
Pillow==7.0.0
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.13
SQLAlchemy-Utils==0.36.1
validators==0.14.2
Werkzeug==1.0.0
Whoosh==2.7.4
WTForms==2.2.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.10.4
